rest_api/serializers.py
from rest_framework import serializers
from .models import Bucketlist


class BucketlistSerializer(serializers.ModelSerializer):
    owner = serializers.ReadOnlyField(source='owner.username')

    class Meta:
        model = Bucketlist
        fields = ('id', 'name', 'owner', 'creation_date', 'modification_date')
        read_only_fields = ('creation_date', 'modification_date')
